requirements.txt
torch==2.5.1+cu121        # Adjust based on CUDA version
torchvision==0.20.1+cu121
transformers==4.40.1
pandas==2.2.3
datasets==3.1.0
matplotlib==3.9.2
accelerate==1.1.0
bitsandbytes==0.44.1
timm==0.4.12
fairscale==0.4.4
pycocoevalcap
Pillow
scipy
clip @ git+https://github.com/openai/CLIP.git
